Add a new 'template' parameter in the POST/send API for test reports
This patch adds a new 'template' parameter in the POST/send API in
order to diversify the use of the test email report feature. So this
gives the ability to choose the email template independently of the
test plan.

By default, if the 'template' parameter is not provided then a
template that matches the test plan name will be used if there is one.
If not then the default 'test.txt' template will be used.

Signed-off-by: Khouloud Touil <ktouil@baylibre.com>
Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>