Add templates.yaml to look up test email report template options
Add templates.yaml as a replacement for the hard-coded Python
dictionary to keep all the per-template options alongside the template
files.  Keep the email template parameter to choose the template and
look up in the YAML config, to still have the flexibility to send
multiple different email reports for a same test plan.  When no
template parameter is provided, look for a template that matches the
test plan name to keep current behaviour.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>