Move the test email template config to templates.yaml
Move the hard-coded configuration of the test email templates to a new
YAML file with the intent to eventually make this part of the rest of
the configuration used in kernelci-core.  It already makes it a bit
easier to add custom template files as no changes in the Python code
are needed any more.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>