Add login test case for baseline plan
When parsing the LAVA callback data for the baseline test plan, add a
"login" test case in the top-level group based on the result of
"auto-login-action" by default.  This is to catch early boot problems
as in the legacy boot test plan, before any test could be run.  Note
that the "login" test case result itself is not LAVA-specific, but
only where it got its status from in this implementation.

As part of this change, stop parsing the "lava" test results for the
"boot" test plan as this is not useful.  The "lava" results are only
used with the "baseline" test plan for the "login" test case now.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>