Create test report summary with dynamic column widths
The test report has some summary tables which can easily get distorted
with long values such as defconfigs with many extra options.  In other
cases, there are some short values but a lot of white space with fixed
widths.  So to solve these issues and make the table scale better,
compute the width of every column dynamically based on the values to
show and squash very long ones with "..." in the middle to not make
the table completely unusable.

Note that it has to be done in Python as it's too much logic for the
Jinja templates.  As a result, the summary tables are passed already
rendered to the template.  Doing this arguably makes the templates
easier to manage anyway, but if it causes some issues then maybe some
custom Jinja extension could be created for dynamic ASCII tables.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>