Use file_server_resource when posting build meta-data
Add file_server_resource to build meta-data POST requests as a
required parameter and use it to determine the location of the build
binaries.  This relies on the client using the same path when first
pushing the build binaries and then posting the build meta-data,
without having the backend to independently reconstruct the path (and
get it wrong in some cases).

This should also help solving issues such as branch names with a slash
or defconfigs with a path to a config fragment.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>