Use file_server_resource when posting build meta-data
Accept the file_server_resource in build POST requests and use it to
determine the location of the build binaries.  This relies on the
client using the same path when pushing the build binaries and then
sending the meta-data, without having the backend to independently
reconstruct the path (and get it wrong in some cases).

This should also help solving issues such as branch names with a slash
or defconfigs with a path to a config fragment.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>